Add files via upload
Welcome to my 2nd Project! In this Project, I useEDA methods and try to predict the professor ratings . As you can guess, there are various methods to suceed this and each method has pros and cons. I think EDA is one of the most important methods because it gives us more insight about the data. In this project we extracted the data from